# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertLifecycleActionRequest": {
        "properties": {
          "expectedVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "expectedVersion"
        ],
        "type": "object"
      },
      "AlertLifecycleResult": {
        "properties": {
          "alertId": {
            "format": "uuid",
            "type": "string"
          },
          "confirmedVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "replayed": {
            "type": "boolean"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "alertId",
          "confirmedVersion",
          "replayed",
          "state"
        ],
        "type": "object"
      },
...
      "CurrentUserResponse": {
        "properties": {
          "developmentAuthentication": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "simulationHandle": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "developmentAuthentication",
          "displayName",
          "organizationId",
          "roles",
          "simulationHandle",
          "userId"
        ],
        "type": "object"
      },
...
      "DevelopmentIdentityResponse": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "simulationHandle": {
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "organizationId",
          "roles",
          "simulationHandle"
        ],
        "type": "object"
      },
      "DevelopmentSessionRequest": {
        "properties": {
          "simulationHandle": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "simulationHandle"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Simulation-only contract. It is not a hospital or clinical API.",
    "title": "Critical Alerts Simulation API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/alerts/{alertId}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "alertId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AlertLifecycleActionRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertLifecycleResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "AlertLifecycleEndpoints"
        ]
      }
    },
...
    "/api/v1/alerts/{alertId}/resolve": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "alertId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AlertLifecycleActionRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertLifecycleResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "AlertLifecycleEndpoints"
        ]
      }
    },
...
    "/api/v1/authorization/administrator": {
      "get": {
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
    "/api/v1/authorization/operator": {
      "get": {
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
    "/api/v1/authorization/organization-scope/{organizationId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "DevelopmentAuthenticationEndpoints"
        ]
      }
    },
    "/api/v1/authorization/practitioner": {
      "get": {
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
...
    "/api/v1/dev/identities": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DevelopmentIdentityResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
...
    "/api/v1/dev/session": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DevelopmentSessionRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
    "/api/v1/dev/session/clear": {
      "post": {
        "responses": {
          "204": {
            "description": "No Content"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "CriticalAlerts.Api"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "DevelopmentAuthenticationEndpoints"
        ]
      }
    },
...
  },
...
}